Agentic/poker_solverllm/requirements.txt
# Main dependencies
openai>=1.0.0
langgraph>=0.0.19
langgraph-prebuilt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-xai>=0.1.0
langchain_core>=0.1.9
pokerkit>=2.0.0
streamlit>=1.32.0
pillow>=10.0.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
pylint>=3.0.0
black>=23.0.0